# src/main/resources/openapi/budgeting.yaml
openapi: 3.0.3
info:
  title: Budgeting API
  version: 1.0.0
servers:
  - url: /
paths:
  /api/v1/budgets:
    post:
      summary: Create a new budget
      operationId: createBudget
      tags:
        - Budgets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBudgetRequest'
      responses:
        '201':
          description: Budget created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
  /api/v1/budgets/{id}:
    get:
      summary: Get a budget by ID
      operationId: getBudget
      tags:
        - Budgets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Budget retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
        '404':
          description: Budget not found
    put:
      summary: Update a budget
      operationId: updateBudget
      tags:
        - Budgets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBudgetRequest'
      responses:
        '200':
          description: Budget updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
        '404':
          description: Budget not found
    delete:
      summary: Delete a budget
      operationId: deleteBudget
      tags:
        - Budgets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Budget deleted successfully
        '404':
          description: Budget not found

components:
  schemas:
    CreateBudgetRequest:
      type: object
      required:
        - organizationalUnitId
        - fiscalYear
        - totalAmount
        - reservedAmount
        - availableAmount
      properties:
        organizationalUnitId:
          type: integer
          format: int64
        fiscalYear:
          type: integer
          format: int32
        totalAmount:
          type: number
        reservedAmount:
          type: number
        availableAmount:
          type: number
    UpdateBudgetRequest:
      type: object
      required:
        - totalAmount
        - reservedAmount
        - availableAmount
      properties:
        totalAmount:
          type: number
        reservedAmount:
          type: number
        availableAmount:
          type: number
    BudgetResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        organizationalUnitId:
          type: integer
          format: int64
        fiscalYear:
          type: integer
          format: int32
        totalAmount:
          type: number
        reservedAmount:
          type: number
        availableAmount:
          type: number
        active:
          type: boolean